refactor: Move to entity command
