+++ lib/core.dart
@@ -9,6 +9,6 @@ export 'src/internal/ioc.dart' show ioc, Service;
 export 'src/constants.dart';
 export 'src/internal/services/http.dart';
 export 'src/internal/managers/event_manager.dart' show Event, Events, MineralEvent;
-export 'src/internal/managers/command_manager.dart' show Command, MineralCommand, Option, OptionType, OptionChoice, Subcommand, CommandGroup;
+export 'src/internal/entities/command.dart' show Command, MineralCommand, Option, OptionType, OptionChoice, Subcommand, CommandGroup;
 export 'src/internal/managers/store_manager.dart' show Store, MineralStore;
 export 'src/internal/managers/module_manager.dart' show Module, MineralModule;
